feat: rework middlewares
