+++ src/rest/rest.application.ts
@@ -5,8 +5,8 @@ import { Component } from '../shared/enums/index.js';
 import { DatabaseClient } from '../shared/libs/database-client/index.js';
 import { getMongoURI } from '../shared/helpers/index.js';
 import express, { Express } from 'express';
-import { Controller, ExceptionFilter } from '../shared/libs/rest/index.js';
-import { ParseTokenMiddleware } from '../shared/libs/rest/middleware/parse-token.middleware.js';
+import { AuthorisationMiddleware, Controller, ExceptionFilter } from '../shared/libs/rest/index.js';
+import { UserService } from '../shared/modules/user/user-service.interface.js';
 
 @injectable()
 export class RestApplication {
@@ -21,6 +21,9 @@ export class RestApplication {
     @inject(Component.CommentController) private readonly commentController: Controller,
     @inject(Component.ExceptionFilter) private readonly appExceptionFilter: ExceptionFilter,
     @inject(Component.AuthExceptionFilter) private readonly authExceptionFilter: ExceptionFilter,
+
+    @inject(Component.UserService) private readonly userService: UserService,
+
   ) {}
 
   private async initExceptionFilters() {
@@ -32,7 +35,7 @@ export class RestApplication {
 
   private async initMiddleware() {
     this.logger.info('Init app-level middleware');
-    const authenticateMiddleware = new ParseTokenMiddleware(this.config.get('JWT_SECRET'));
+    const authenticateMiddleware = new AuthorisationMiddleware(this.config.get('JWT_SECRET'), this.userService);
 
     this.server.use(express.json());
     this.server.use(
