+++ src/shared/libs/rest/index.ts
@@ -14,7 +14,6 @@ export * from './middleware/validate-dto.middleware.js';
 export * from './middleware/document-exists.middleware.js';
 export * from './middleware/unique-email.middleware.js';
 export * from './middleware/upload-file.middleware.js';
-export * from './middleware/parse-token.middleware.js';
-export * from './middleware/private-route.middleware.js';
 export * from './middleware/authorisation.middleware.js';
+export * from './middleware/private-route.middleware.js';
 export * from './middleware/author.middleware.js';
